refactor: change directory to entity from model
