+++ src/main/kotlin/com/group4/ticketingservice/entity/Bookmark.kt
@@ -1,4 +1,4 @@
-package com.group4.ticketingservice.model
+package com.group4.ticketingservice.entity
 
 import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
