+++ src/main/kotlin/com/group4/ticketingservice/repository/BookmarkRepository.kt
@@ -1,6 +1,6 @@
 package com.group4.ticketingservice.repository
 
-import com.group4.ticketingservice.model.Bookmark
+import com.group4.ticketingservice.entity.Bookmark
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
