+++ src/test/kotlin/com/group4/ticketingservice/bookmark/BookmarkControllerTest.kt
@@ -2,7 +2,7 @@ package com.group4.ticketingservice.bookmark
 
 import com.group4.ticketingservice.controller.BookmarkController
 import com.group4.ticketingservice.dto.BookmarkFromdto
-import com.group4.ticketingservice.model.Bookmark
+import com.group4.ticketingservice.entity.Bookmark
 import com.group4.ticketingservice.service.BookmarkService
 import com.ninjasquad.springmockk.MockkBean
 import io.mockk.every
